+++ src/test/java/picocli/UnmatchedOptionTest.java
@@ -1,6 +1,5 @@
 package picocli;
 
-import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.contrib.java.lang.system.ProvideSystemProperty;
@@ -11,6 +10,9 @@ import picocli.CommandLine.Option;
 import picocli.CommandLine.OverwrittenOptionException;
 import picocli.CommandLine.Parameters;
 import picocli.CommandLine.UnmatchedArgumentException;
+
+import java.util.Arrays;
+import java.util.List;
 
 import static org.junit.Assert.*;
 
@@ -203,7 +205,7 @@ public class UnmatchedOptionTest {
     // -----------------------------------
 
     @Test
-    public void testSingleValuePositionalDoesNotConsumeArgResemblingOption() {
+    public void testSingleValuePositionalDoesNotConsumeUnknownOption() {
         class App {
             @Option(names = "-x") int x;
             @Parameters String y;
@@ -213,7 +215,7 @@ public class UnmatchedOptionTest {
         expect(new App(), "Unknown option: '-z'", UnmatchedArgumentException.class, "-x", "3", "-z", "4");
     }
     @Test
-    public void testSingleValuePositionalCanBeConfiguredToConsumeArgResemblingOption() {
+    public void testSingleValuePositionalCanBeConfiguredToConsumeUnknownOption() {
         class App {
 
             @Option(names = "-x") int x;
@@ -230,7 +232,7 @@ public class UnmatchedOptionTest {
     }
 
     @Test
-    public void testMultiValuePositionalDoesNotConsumeArgResemblingOption() {
+    public void testMultiValuePositionalDoesNotConsumeUnknownOption() {
         class App {
             @Option(names = "-x") int x;
             @Parameters String[] y;
@@ -241,7 +243,7 @@ public class UnmatchedOptionTest {
         expect(new App(), "Unknown option: '-z'", UnmatchedArgumentException.class, "-x", "3", "4", "-z");
     }
     @Test
-    public void testMultiValuePositionalCanBeConfiguredToConsumeArgResemblingOption() {
+    public void testMultiValuePositionalCanBeConfiguredToConsumeUnknownOption() {
         class App {
             @Option(names = "-x") int x;
             @Parameters String[] y;
@@ -264,7 +266,7 @@ public class UnmatchedOptionTest {
     }
 
     @Test
-    public void testMultiValueVarArgPositionalDoesNotConsumeArgResemblingOptionSimple() {
+    public void testMultiValueVarArgPositionalDoesNotConsumeUnknownOptionSimple() {
         class App {
             @Parameters(arity = "*") String[] y;
         }
@@ -274,7 +276,7 @@ public class UnmatchedOptionTest {
     }
 
     @Test
-    public void testMultiValueVarArgPositionalDoesNotConsumeArgResemblingOption() {
+    public void testMultiValueVarArgPositionalDoesNotConsumeUnknownOption() {
         class App {
             @Option(names = "-x") int x;
             @Parameters(arity = "*") String[] y;
@@ -286,7 +288,7 @@ public class UnmatchedOptionTest {
         expect(new App(), "Unknown option: '-z'", UnmatchedArgumentException.class, "-x", "3", "4", "-z");
     }
     @Test
-    public void testMultiValueVarArgPositionalCanBeConfiguredToConsumeArgResemblingOption() {
+    public void testMultiValueVarArgPositionalCanBeConfiguredToConsumeUnknownOption() {
         class App {
             @Option(names = "-x") int x;
             @Parameters(arity = "*") String[] y;
@@ -309,7 +311,7 @@ public class UnmatchedOptionTest {
     }
 
     @Test
-    public void testMultiValuePositionalArity2_NDoesNotConsumeArgResemblingOption() {
+    public void testMultiValuePositionalArity2_NDoesNotConsumeUnknownOption() {
         class App {
             @Option(names = "-x") int x;
             @Parameters(arity = "2..*") String[] y;
@@ -324,7 +326,7 @@ public class UnmatchedOptionTest {
         expect(new App(), "Unknown option: '-z'", UnmatchedArgumentException.class, "-x", "3", "4", "-z");
     }
     @Test
-    public void testMultiValuePositionalArity2_NCanBeConfiguredToConsumeArgResemblingOption() {
+    public void testMultiValuePositionalArity2_NCanBeConfiguredToConsumeUnknownOption() {
         class App {
             @Option(names = "-x") int x;
             @Parameters(arity = "2..*") String[] y;
@@ -344,78 +346,238 @@ public class UnmatchedOptionTest {
         assertArrayEquals(new String[]{"4", "-z"}, app.y);
     }
 
-    @Ignore
     @Test
-    public void testSingleValueOptionDoesNotConsumeArgResemblingOption() {
+    public void testSingleValueOptionDoesNotConsumeUnknownOptionByDefault() {
         class App {
             @Option(names = "-x") int x;
             @Option(names = "-y") String y;
         }
 
-        //FIXME
-        expect(new App(), "Unknown option: '-z'; Missing required parameter for option '-y' (<y>)", MissingParameterException.class, "-y", "-z");
+        CommandLine cmd = new CommandLine(new App());
+        cmd.setUnmatchedOptionsAllowedAsOptionParameters(false);
+        expect(cmd, "Unknown option: '-z'; Expected parameter for option '-y' but found '-z'", UnmatchedArgumentException.class, "-y", "-z");
     }
 
-    @Ignore
     @Test
-    public void testMultiValueOptionDoesNotConsumeArgResemblingOption() {
+    public void testSingleValueOptionCanBeConfiguredToConsumeUnknownOption() {
+        class App {
+            @Option(names = "-x") int x;
+            @Option(names = "-y") String y;
+        }
+
+        App app = new App();
+        new CommandLine(app).setUnmatchedOptionsAllowedAsOptionParameters(true).parseArgs("-y", "-z");
+        assertEquals("-z", app.y);
+    }
+
+    @Test
+    public void testMultiValueOptionDoesNotConsumeUnknownOptionByDefault() {
         class App {
             @Option(names = "-x") int x;
             @Option(names = "-y") String[] y;
         }
 
-        //FIXME
-        expect(new App(), "Unknown option: '-z'; Missing required parameter for option '-y' (<y>)", MissingParameterException.class, "-y", "-z");
-        expect(new App(), "Unknown option: '-z'", UnmatchedArgumentException.class, "-y", "3", "-z");
+        CommandLine cmd = new CommandLine(new App());
+        cmd.setUnmatchedOptionsAllowedAsOptionParameters(false);
+        expect(cmd, "Unknown option: '-z'; Expected parameter for option '-y' but found '-z'", UnmatchedArgumentException.class, "-y", "-z");
+        expect(cmd, "Unknown option: '-z'", UnmatchedArgumentException.class, "-y", "3", "-z");
     }
 
-    @Ignore
     @Test
-    public void testMultiValueOptionArity2DoesNotConsumeArgResemblingOption() {
+    public void testMultiValueOptionCanBeConfiguredToConsumeUnknownOption() {
+        class App {
+            @Option(names = "-x") int x;
+            @Option(names = "-y") String[] y;
+        }
+
+        App app = new App();
+        new CommandLine(app).setUnmatchedOptionsAllowedAsOptionParameters(true).parseArgs("-y", "-z");
+        assertArrayEquals(new String[]{"-z"}, app.y);
+
+        // arity=1 for multi-value options
+        CommandLine cmd = new CommandLine(new App()).setUnmatchedOptionsAllowedAsOptionParameters(true);
+        expect(cmd, "Unknown option: '-z'", UnmatchedArgumentException.class, "-y", "3", "-z");
+    }
+
+    @Test
+    public void testMultiValueOptionArity2DoesNotConsumeUnknownOptionByDefault() {
         class App {
             @Option(names = "-x") int x;
             @Option(names = "-y", arity = "2") String[] y;
         }
 
-        expect(new App(), "option '-y' at index 0 (<y>) requires at least 2 values, but only 1 were specified: [-z]",
+        CommandLine cmd = new CommandLine(new App());
+        expect(cmd, "option '-y' at index 0 (<y>) requires at least 2 values, but only 1 were specified: [-z]",
                 MissingParameterException.class, "-y", "-z");
 
-        //FIXME
-        expect(new App(), "Expected parameter 2 (of 2 mandatory parameters) for option '-y' but found '-z'",
-                MissingParameterException.class, "-y", "3", "-z");
+        cmd.setUnmatchedOptionsAllowedAsOptionParameters(false);
+        expect(cmd, "Unknown option: '-z'; Expected parameter 2 (of 2 mandatory parameters) for option '-y' but found '-z'",
+                UnmatchedArgumentException.class, "-y", "3", "-z");
     }
 
-    @Ignore
     @Test
-    public void testMultiValueOptionArity2_NDoesNotConsumeArgResemblingOption() {
+    public void testMultiValueOptionArity2CanBeConfiguredToConsumeUnknownOptions() {
+        class App {
+            @Option(names = "-x") int x;
+            @Option(names = "-y", arity = "2") String[] y;
+        }
+
+        CommandLine cmd = new CommandLine(new App()).setUnmatchedOptionsAllowedAsOptionParameters(true);
+        expect(cmd, "option '-y' at index 0 (<y>) requires at least 2 values, but only 1 were specified: [-z]",
+                MissingParameterException.class, "-y", "-z");
+
+        App app = new App();
+        new CommandLine(app).setUnmatchedOptionsAllowedAsOptionParameters(true).parseArgs("-y", "3", "-z");
+        assertArrayEquals(new String[]{"3", "-z"}, app.y);
+    }
+
+    @Test
+    public void testMultiValueOptionArity2_NDoesNotConsumeUnknownOptionsByDefault() {
         class App {
             @Option(names = "-x") int x;
             @Option(names = "-y", arity = "2..N") String[] y;
         }
 
-        expect(new App(), "option '-y' at index 0 (<y>) requires at least 2 values, but only 1 were specified: [-z]",
+        CommandLine cmd = new CommandLine(new App());
+        cmd.setUnmatchedOptionsAllowedAsOptionParameters(false);
+        expect(cmd, "option '-y' at index 0 (<y>) requires at least 2 values, but only 1 were specified: [-z]",
                 MissingParameterException.class, "-y", "-z");
 
-        //FIXME
-        expect(new App(), "Expected parameter 2 (of 2 mandatory parameters) for option '-y' but found '-z'",
-                MissingParameterException.class, "-y", "1", "-z");
+        expect(cmd, "Unknown option: '-z'; Expected parameter 2 (of 2 mandatory parameters) for option '-y' but found '-z'",
+                UnmatchedArgumentException.class, "-y", "1", "-z");
 
-        //FIXME
-        expect(new App(), "Unknown option: '-z'", UnmatchedArgumentException.class, "-y", "1", "2", "-z");
+        expect(cmd, "Unknown option: '-z'; Expected parameter 3 (of 2 mandatory parameters) for option '-y' but found '-z'",
+                UnmatchedArgumentException.class, "-y", "1", "2", "-z");
     }
 
-
-    @Ignore
     @Test
-    public void testMultiValueVarArgOptionDoesNotConsumeArgResemblingOption() {
+    public void testMultiValueOptionArity2_NCanBeConfiguredToConsumeUnknownOption() {
+        class App {
+            @Option(names = "-x") int x;
+            @Option(names = "-y", arity = "2..N") String[] y;
+        }
+
+        CommandLine cmd = new CommandLine(new App()).setUnmatchedOptionsAllowedAsOptionParameters(true);
+        expect(cmd, "option '-y' at index 0 (<y>) requires at least 2 values, but only 1 were specified: [-z]",
+                MissingParameterException.class, "-y", "-z");
+
+        App app = new App();
+        new CommandLine(app).setUnmatchedOptionsAllowedAsOptionParameters(true).parseArgs("-y", "1", "-z");
+        assertArrayEquals(new String[]{"1", "-z"}, app.y);
+
+        app = new App();
+        new CommandLine(app).setUnmatchedOptionsAllowedAsOptionParameters(true).parseArgs("-y", "1", "2", "-z");
+        assertArrayEquals(new String[]{"1", "2", "-z"}, app.y);
+    }
+
+    @Test
+    public void testMultiValueVarArgOptionDoesNotConsumeUnknownOptionsByDefault() {
         class App {
             @Option(names = "-x") int x;
             @Option(names = "-y", arity = "*") String[] y;
         }
 
-        //FIXME
-        //expect(new App(), "Missing required parameter for option '-y' (<y>)", MissingParameterException.class, "-y", "-z");
-        expect(new App(), "Unknown option: '-z'", UnmatchedArgumentException.class, "-y", "3", "-z");
+        CommandLine cmd = new CommandLine(new App());
+        cmd.setUnmatchedOptionsAllowedAsOptionParameters(false);
+        expect(cmd, "Unknown option: '-z'; Expected parameter for option '-y' but found '-z'", UnmatchedArgumentException.class, "-y", "-z");
+        expect(cmd, "Unknown option: '-z'; Expected parameter for option '-y' but found '-z'", UnmatchedArgumentException.class, "-y", "3", "-z");
     }
 
+    @Test
+    public void testMultiValueVarArgOptionCanBeConfiguredToConsumeUnknownOption() {
+        class App {
+            @Option(names = "-x") int x;
+            @Option(names = "-y", arity = "*") String[] y;
+        }
+
+        App app = new App();
+        new CommandLine(app).setUnmatchedOptionsAllowedAsOptionParameters(true).parseArgs("-y", "-z");
+        assertArrayEquals(new String[]{"-z"}, app.y);
+
+        app = new App();
+        new CommandLine(app).setUnmatchedOptionsAllowedAsOptionParameters(true).parseArgs("-y", "3", "-z");
+        assertArrayEquals(new String[]{"3", "-z"}, app.y);
+    }
+
+    @Test //#639
+    public void testUnknownOptionAsOptionValue() {
+        class App {
+            @Option(names = {"-x", "--xvalue"})
+            String x;
+
+            @Option(names = {"-y", "--yvalues"}, arity = "1")
+            List<String> y;
+        }
+        App app = new App();
+        CommandLine cmd = new CommandLine(app).setUnmatchedOptionsAllowedAsOptionParameters(true);
+        cmd.parseArgs("-x", "-unknown");
+        assertEquals("-unknown", app.x);
+
+        cmd.parseArgs("-y", "-unknown");
+        assertEquals(Arrays.asList("-unknown"), app.y);
+
+        cmd = new CommandLine(new App());
+        cmd.setUnmatchedOptionsAllowedAsOptionParameters(false);
+        expect(cmd, "Unknown option: '-unknown'; Expected parameter for option '--xvalue' but found '-unknown'", UnmatchedArgumentException.class, "-x", "-unknown");
+        expect(cmd, "Unknown option: '-unknown'; Expected parameter for option '--yvalues' but found '-unknown'", UnmatchedArgumentException.class, "-y", "-unknown");
+    }
+
+    static class LenientShortConverter implements CommandLine.ITypeConverter<Short> {
+        public Short convert(String value) {
+            return Short.decode(value); // allow octal values "-#123"
+        }
+    }
+    static class LenientLongConverter implements CommandLine.ITypeConverter<Long> {
+        public Long convert(String value) {
+            return Long.decode(value); // allow hex values "-0xABCDEF"
+        }
+    }
+    @Test //#639
+    public void testNegativeNumbersAreNotUnknownOption() {
+        class App {
+            @Option(names = {"-i", "--int"})
+            int i;
+
+            @Option(names = {"-s", "--short"}, converter = LenientShortConverter.class)
+            short s;
+
+            @Option(names = {"-L", "--long"}, converter = LenientLongConverter.class)
+            long L;
+
+            @Option(names = {"-d", "--double"})
+            double d;
+
+            @Option(names = {"-f", "--float"})
+            float f;
+
+            @Option(names = {"-x", "--xvalue"})
+            String x;
+
+            @Option(names = {"-y", "--yvalues"}, arity = "1")
+            List<String> y;
+        }
+
+        App app = new App();
+        CommandLine cmd = new CommandLine(app).setUnmatchedOptionsAllowedAsOptionParameters(true);
+        String[] args = {"-i", "-1", "-s", "-#55", "-L", "-0xCAFEBABE", "-d", "-Infinity", "-f", "-NaN", "-x", "-2", "-y", "-3", "-y", "-0", "-y", "-0.0", "-y", "-NaN"};
+        cmd.parseArgs(args);
+        assertEquals(-1, app.i);
+        assertEquals(-0x55, app.s);
+        assertEquals(-0xCAFEBABEL, app.L);
+        assertEquals(Double.NEGATIVE_INFINITY, app.d, 0);
+        assertEquals(Float.NaN, app.f, 0);
+        assertEquals("-2", app.x);
+        assertEquals(Arrays.asList("-3", "-0", "-0.0", "-NaN"), app.y);
+
+        cmd = new CommandLine(new App());
+        cmd.setUnmatchedOptionsAllowedAsOptionParameters(false);
+        cmd.parseArgs(args);
+        assertEquals(-1, app.i);
+        assertEquals(-0x55, app.s);
+        assertEquals(-0xCAFEBABEL, app.L);
+        assertEquals(Double.NEGATIVE_INFINITY, app.d, 0);
+        assertEquals(Float.NaN, app.f, 0);
+        assertEquals("-2", app.x);
+        assertEquals(Arrays.asList("-3", "-0", "-0.0", "-NaN"), app.y);
+    }
 }